petrolprice/app/requirements.txt
requests>=2.28.0,<3
Pillow>=10.0.0
pytesseract>=0.3.10
aiohttp>=3.9.0
